First fully working thread-safe version
Former-commit-id: 4a7ddb80b94e46458a692b8845bd2faac56bf4b3

+++ Tracking/TrkVertexFitter/TrkVKalVrtFitter/src/TrkCascadeFitter.cxx
@@ -316,7 +316,7 @@ VxCascadeInfo * TrkVKalVrtFitter::fitCascade(const Vertex* primVrt, bool FirstDe
 //
     if(!m_isFieldInitialized)setInitializedField();    //to allow callback for init
 //
-    long int ntrk=0;
+    int ntrk=0;
     StatusCode sc; sc.setChecked();
     std::vector<const TrackParameters*> baseInpTrk;
     if(m_firstMeasuredPoint){               //First measured point strategy
